Fix warnings; JIRA:RW-2061
git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@100348 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ include/objtools/readers/agp_util.hpp
@@ -836,7 +836,7 @@ public:
     // 0: reading from STDIN or from a single file
     int GetFileNum()
     {
-        return m_InputFiles.size();
+        return static_cast<int>(m_InputFiles.size());
     }
 
     const string& GetFile(int num)
